do not use global variables

+++ js/modules/Debug/Debug.js
@@ -328,7 +328,7 @@ window.GLPI.Debug = new class Debug {
             // update the total counters
             data.forEach((query) => {
                 sql_data.total_requests += 1;
-                sql_data.total_duration += parseFloat(query['time']);
+                sql_data.total_duration += parseInt(query['time']);
             });
         });
